minor refactoring and travis badge
