+++ README.md
@@ -1,2 +1,5 @@
 # Collagen
+[![Build Status](https://travis-ci.com/MIPT-Oulu/Collagen.svg?branch=master)](https://travis-ci.com/MIPT-Oulu/Collagen)
+
+
 Deep Learning Framework for Reproducible Experiments
